media/gpu/test: Increase event timeout in video decoder (performance) tests.
This CL increases the timeout used when waiting for events in the
video_decode_accelerator_perf_tests. The MeasureCappedPerformance test decodes
the video in real-time, so we should make sure the timeout is at least as large
as the video's duration.

The timeout is also increased when writing video frames to disk in the
video_decode_accelerator_tests, as this can take a lot of time.

TEST=./video_decode_accelerator_tests on eve

BUG=None

Change-Id: Iaf92d32cae755e6d0afe50571a87bb44a87c561c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1731641
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#683933}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 458351fd7b4f67b9900686872e73ac9a4d7e63f1