media/gpu/v4l2svda: use queue of GL fences
As we are planning to add IP support, GL fences will not always be
associated to decoder buffers - if the IP is in use, then the fence will
refer to the IP buffer and we must thus prevent it from being used for
processing again.

However, the way GL fences management is done is to wait on a fence
before queuing a decoder buffer. This CL addresses this by using a FIFO
queue of GL fences which retain their associated decode surface with
them. Since the IP buffer will be tied to the decode surface, this means
it will be kept for as long as required.

This design is similar to what was done in the V4L2VDA to address the
same situation.

Bug: b:132589320
Test: vdaunittest and vdatests pass on Minnie.
Change-Id: I06ba55d995a5ba3bc9fc09fcc4b2358987294eb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1741431
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#685456}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6ddfe0d85b90f8e74d6a77f14ee980288b51b047