media/gpu/v4l2svda: fix frame referencing with request API
So far, we used the OUTPUT buffer id as the timestamp for referencing a
frame. However, this method is flawed: the input buffer can potentially
be reused as soon as the driver returns it, meaning the same reference
ID can be reused before the driver is done referencing the previously
rendered frame. Doing so results in corruption of the following frames.

Fix this by using the CAPTURE buffer id instead: since user-space is
responsible for not resubmitting frames until the decoder is done
referencing them, this method is safe and ensures a timestamp still in
use is never re-submitted.

Bug: 993695
Test: VDA Thumbnails test passes on Kukui.

Change-Id: I977b1f06f34e1cfa18f6209979cda0394c0c5bfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1752578
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#686768}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb4ba0d8eaf7f6abfbaea8a6e446345ae461ad81