Reland "Use REALTIME_AUDIO thread for Fake Input Stream"
This is a reland of ae18b126269f138d10067ceb35e5876cf55054fd

The problem with the first CL was that calling FakeAudioInputStream::Stop()
calls Stop() on the capture thread. Thread::Stop() triggers a join,
which blocks. On Mac, the AudioThread is actually the browser UI thread,
and this block is not ok.

The fix to mitigate this issue is to have the capture thread owned in
a std::unique ptr that deletes the capture thread on the AudioManager
worker thread. The capture thread will be stopped on the worker thread
when it is deleted.

Original change's description:
> Use REALTIME_AUDIO thread for Fake Input Stream
>
> Delays were seen when using the Audio Thread, detailed in the attached
> bug. The use of the REALTIME_AUDIO priority fixes these delays.
>
> Bug: 971265
> Change-Id: Ieaf280b966354e3515f038ab81841572e075f0b8
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1655530
> Reviewed-by: Max Morin <maxmorin@chromium.org>
> Commit-Queue: Evan Shrubsole <eshr@google.com>
> Cr-Commit-Position: refs/heads/master@{#670443}

Bug: 971265
Change-Id: I6e794042fe040ad87ffe823789666c5a987ce9cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1702426
Commit-Queue: Evan Shrubsole <eshr@google.com>
Reviewed-by: Henrik Andreasson <henrika@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#686796}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 75d6e4fbd96965b0f58674f0d26339ceb654a9b4