Move CodecBufferWaitCoordinator ownership to CodecSurfaceBundle.
CodecSurfaceBundle now owns CodecBufferWaitCoordinator instead of
TextureOwner. This is because when switching codecs, we re-use same
TextureOwner to create a new CodecBufferWaitCorrdinator in
VideoFrameFactoryImpl which was an issue. This could lead to multiple
CodecBufferWaitCoordinator to be in flight at once.
Hence instead of creating CodecBufferWaitCoordinator every time we
switch codecs, we just use the one which CodecSurfaceBundle has.

Bug: 996382
Change-Id: I614a2dbb54203f397dbbb9c1d6fa772e4b0146d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1764412
Commit-Queue: vikas soni <vikassoni@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#689241}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0ac92ad68e675c7b5371ed10e49705d81c7644e0