Fix for failing 4-channel microphone not working properly.
This is a minor change on how the audio format is setup for low-latency
input streams on Windows in combination with a microphone with 4 channels.
Previous changes have resolved mutli-channel issues in combination with
DISCRETE channel layout but this particular device reports QUAD and we
must therefore change how the channel mask parameter is acquired.

Before this patch we asked the device and it said QUAD but that does not
match the fact that the core audio layer forces number of channels to 2 for
multi-channel cases. Hence, we had a conflict of 2 channels and QUAD.

With the fix below, the final format will contain 2 channels and STEREO and
it resolves the issue.

In summary: avoid asking the device about it channel format since it can
conflict with what the CoreAudioUtil class has set.

The change has been verified by a reporting user and I have tested it on
other multi-channel devices as well.

Bug: 933976
Change-Id: I3b566afbd35bcf3afb9684b81bf06ee9bd4daded
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1772877
Commit-Queue: Henrik Andreasson <henrika@chromium.org>
Reviewed-by: Henrik Grunell <grunell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#691127}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 584f7ed6cba43863fa97b549e41dbe545ddab57c