media/gpu/v4l2: Fix timestamps not always being cleared in V4L2SliceVideoDecoder
For each decode request in the V4L2SliceVideoDecoder the time at which the
decode operation started is recorded in a map. This entry is removed again when
a surface is outputted with associated buffer ID. Unfortunately not each decode
operation will lead to a surface being output, so the map will keep on growing
during playback. This CL fixes this by making use of a cache instead of a map.

TEST=./video_decode_accelerator_tests on Kevin

BUG=chromium:997096,b/140287647

Change-Id: I1ed78e51218d15d57fe487df02b58b0652e1ae42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1777845
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#694125}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 59766d629973a916e662a8726872ba6f163abb44