[CoreAudio] Update core audio work around for Catalina.
Recently, UMA stats indicated that the work around set in place to make
sure that PauseIO and ResumeIO are executed sequentially stopped
working. The suspicion was that the CoreAudio internals changed in such
a way that the the targeted code is now called from a different
location.

This CL fixes the hack for Catalina, so that the symbols searched
include the ones now used in Catalina. I have confirmed that these names
are used by running a custom build of chrome and later checking that in
chrome://histograms the events of pause and resume have been recorded.

BUG=976204

Change-Id: If137d327c2415b95ad0211a5fb9bfe4930525816
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1803435
Reviewed-by: Tommi <tommi@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Oskar Sundbom <ossu@chromium.org>
Commit-Queue: Armando Miraglia <armax@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#696434}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 05511c7776b404c2936332dfcfba221843526c80