If the backend texture is the same we let skia take care of it
We are now using SharedImage to hold the image for the video renderer,
https://chromium-review.googlesource.com/c/chromium/src/+/1616978.
That change, together with the change that made the video be deferred, and
therefore, use the paintRecorder, let us safely remove the workaround.

There was nevertheless the issue of texture lifetime to extend the texture.
If the backend texture is the same in the SkImage and the one contained
in the cache, we will let skia release it. If they are different, cache has
to destroy their own texture, and the one in skia will survive the delete.

Bug: 974656, 991792, 993774, 994516, 996083, 1004780
Change-Id: I3bf8c1037b77434a1cf0385872fc7d994886cd30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1789883
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Khushal <khushalsagar@chromium.org>
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#697370}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d5ce90c8935ba2887a423cc69af53d710d9c2220