Provide LearningTask to MojoLearningTaskController.
It turns out to be very convient to have the LearningTask on the
client side of the mojo connection.  This CL allows us to specify a
LearningTask during construction of the client, that is returned
by GetLearningTask.

Change-Id: I41e852f338dcbeed26a2a20022e8ceaa5226a649
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1809891
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#697702}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f050573fce8cb08b662a6fcef3b074c334fae89e