Reland "Pool SharedImageVideo for MCVD."
This re-lands:

https://chromium-review.googlesource.com/c/chromium/src/+/1699138

which was reverted by:

https://chromium-review.googlesource.com/c/chromium/src/+/1824081

due to asan failures.

The failures were test-only.  It used a base::Unretained in a mock
that, due to a posted callback, was being destroyed during test
teardown before the callback had a chance to run.  When it did run,
asan got mad.

This CL replaces the unretained with a weak ptr.  Changes are in
mock_shared_image_video_provider.* .

TBR=sandersd,vikassoni,mpearson

Change-Id: I85e6f85a6f81964a6c4b286d2b8219d0b4840958
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1825900
Commit-Queue: Frank Liberato <liberato@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#699994}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f502df718cb143a7183c16b8d27838cb0bd253cb