media/gpu/vaapi: Specify the surface usage hint in vaCreateSurface()
This specifies a surface usage hint in vaCreateSurfaces(). i965 driver
doesn't change a behavior by this change. The alignment of a height of
a created surface is changed to 64 to 1 with iHD driver. Thanks to this,
the drivers' alignment rule is or equal to weaker than the minigbm
alignment rule on encoding.

Bug: 1006151
Test: VEA unittest, VDA unittest, JDA unittest
Change-Id: I4b0d0bd58c7c7f39151bc91b164b1f5aaf7c63bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1816479
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#700108}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 219fbd9c64a0ad1f775748a8ab06c53ddd082bb6