vaapi: Improve VAAPI encoding tracing
While investigating if we have any lock contention in VaapiWrapper we
noticed in VaapiVideoEncodeAccelerator many parts of the decoding are
not taken into account.
This CL adds traces on parts that accounts for the total encode time
and that were not traced already.

in particular:
 - VaapiWrapper::DownloadFromVABuffer that is the time it takes to
   copy the encoded stream into shared memory.
 - VaapiWrapper::CreateVABuffer, needed to create a buffer to decode
   into, while it's fast by itself, this might be waiting on another
   encode to finish.
 - VaapiWrapper::DestroyPendingBuffers, that is called before
   dowloading the buffer and it's in the critical path.
 - VAVEA::FromExecuteToReturn, it's the time from VAAPI::Execute to when
   we call the callback to notify the encode has finished.
 - VAVEA::FromCreateEncodeJobToReturn, is the time from when we start trying
   to process the encoding to when we notify the client the encode
   is done.

We additionally added a few "Locked" traces, like for example

VaapiWrapper::Execute and VaapiWrapper::ExecuteLocked, with the second
one starting tracing after the Lock is acquired, in this way it's easy
to see where we're waiting for vaapi and where we're waiting to acquire
a lock.

https://screenshot.googleplex.com/maVPXzKd7v0 makes clear how we were not
accounting for all the decoding. VAVEA:: Execute trackes only a tiny
percentange of the decoding wall time and DownloadFromVABuffer might take
significantly longer.

Bug: 1005592
Test: Trace a video a encode and look at it.

Change-Id: Ic7e5d58442238241b573d05c4ad86c2443aed74d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1815951
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#701007}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9378b5a802a413793f21b419ccd0eb5dcf693d54