media: Retain reference to the wrapped frame in WrapVideoFrame
When we use VideoFrame::WrapVideoFrame to soft-apply a different set
of format, visible rectangle, and natural size on a source VideoFrame,
we have to make sure the source VideoFrame outlives the wrapping frame
in most if not all the scenarios.

This CL explicitly retains a reference to the wrapped frame inside the
wrapping frame when WrapVideoFrame is called.  Along with the change,
we can remove several empty destruction callbacks that are only used
for retaining a reference to the wrapped frame.

Bug: 982201
Change-Id: Id81ad65c500fe4f0b47a60d98819408522314129
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1830241
Commit-Queue: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#701466}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b2075eb504d37a2b38fc53bff8156b5b18058bde