Do not add destruction observer to PooledSharedImageVideoProvider
PooledSharedImageVideoProvider::GpuHelperImpl currently registers
a stub destruction observer in its constructor. It does not remove
this observer in its destructor, which can lead to a crash when
CommandBufferStub attempts to notify it of destruction.

I was going to unregister in destruction, but then noticed that
this observer isn't needed:
- CommandBufferHelper already registers its own observer internally
- GpuHelperImpl just used this notification to clear
  |command_buffer_helper_|, which is already taken care of internally
  in that class.

So instead this CL just removes the observer from GpuHelperImpl.

Bug: 1010612
Change-Id: I47d1982d6ccaf8624650a4274e8cb1500aba4f9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1835168
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Eric Karl <ericrk@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#702204}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 858b4170af50bb23cb1583e3c5e241c66f41f127