media/gpu/v4l2: add request API support for VD
This patch is part of the first phase for adding request API support to
VD.
The implementation is similar to the implementation in VDA.

BUG=chromium:1009921
TEST=ran video_decode_accelerator_tests on veyron_minnie-kernelnext with
--use_vd option and checked the log contains "Using request API" message
which indicates request API are used on being veyron_minnie-kernelnext.
Also confirmed the tests pass without the VD option.

Signed-off-by: Francois Buergisser <fbuergisser@chromium.org>
Change-Id: I05d0a6221e7c1044baf31cec70d27f4685ffb255
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1833343
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#702362}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7fee4d11a8f100f68310bdc41c4849ad97452bc5