MSE: Make SBStream::Append return void
Long ago, this method could return false if internal buffering detected
a problem in the buffer sequence/metadata being appended. With the
MseBufferByPts-related cleanup, this method was made to always return
true, with a note that it might in future return false if one current
CHECK is changed to a DCHECK. This CL asserts that that CHECK is
necessary to verify a key assumption hasn't regressed in release
builds, and changes the interface to void to match current logic and
usage.

BUG=580621

Change-Id: Ie81d6f1681b65ca93e6af1f21d6f2289c6c0b7da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1841237
Commit-Queue: Matthew Wolenetz <wolenetz@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#703067}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8151b15cbec30a8bed834eef102e5933fc176f2c